fix: cast libc lock constants to i16 for newer Linux glibc
On newer Linux systems, libc::F_WRLCK and libc::F_UNLCK are i32 but
the flock struct's l_type field is i16. Add explicit casts to fix
the type mismatch errors.

+++ src/ida/lock.rs
@@ -270,7 +270,7 @@ fn try_lock_file(file: &File) -> Result<(), u32> {
     use std::os::unix::io::AsRawFd;
 
     let mut fl = libc::flock {
-        l_type: libc::F_WRLCK,
+        l_type: libc::F_WRLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -304,7 +304,7 @@ fn locked_by_pid(path: &Path) -> Option<u32> {
         .ok()?;
 
     let mut fl = libc::flock {
-        l_type: libc::F_WRLCK,
+        l_type: libc::F_WRLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -318,7 +318,7 @@ fn locked_by_pid(path: &Path) -> Option<u32> {
     if rc == -1 {
         return None;
     }
-    if fl.l_type == libc::F_UNLCK {
+    if fl.l_type == libc::F_UNLCK as i16 {
         None
     } else {
         Some(fl.l_pid as u32)
@@ -335,7 +335,7 @@ fn locked_by_pid_from_fd(file: &File) -> Option<u32> {
     use std::os::unix::io::AsRawFd;
 
     let mut fl = libc::flock {
-        l_type: libc::F_WRLCK,
+        l_type: libc::F_WRLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -348,7 +348,7 @@ fn locked_by_pid_from_fd(file: &File) -> Option<u32> {
     if rc == -1 {
         return None;
     }
-    if fl.l_type == libc::F_UNLCK {
+    if fl.l_type == libc::F_UNLCK as i16 {
         None
     } else {
         Some(fl.l_pid as u32)
